% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-methods.R
\name{as_SurvivalProbabilities.CDSCurve}
\alias{as_SurvivalProbabilities.CDSCurve}
\title{Bootstraps Survival Probabilitie from a CDS curve
Using \href{https://www.rdocumentation.org/packages/credule/versions/0.1.3}{credule package.}
The output of bootstrapping is a vector of cumulative survival probabilities.}
\usage{
\method{as_SurvivalProbabilities}{CDSCurve}(x, zero_curve,
  num_timesteps_pa = 12, accrued_premium = TRUE)
}
\arguments{
\item{x}{An object of type \code{CDSCurve}}

\item{zero_curve}{An object of type \code{ZeroCurve}}

\item{num_timesteps_pa}{It represents the number of timesteps used to perform the numerical integral
required while computing the default leg value. It is shown that a monthly discretisation
usually gives a good approximation (Ref. Valuation of Credit Default Swaps, Dominic O Kane and
Stuart Turnbull)}

\item{accrued_premium}{If set to TRUE, the accrued premium will be taken into account in the calculation of the premium leg value.}
}
\value{
An object of type \code{SurvivalProbabilitiesCurve}
}
\description{
Bootstraps Survival Probabilitie from a CDS curve
Using \href{https://www.rdocumentation.org/packages/credule/versions/0.1.3}{credule package.}
The output of bootstrapping is a vector of cumulative survival probabilities.
}
