% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-methods.R
\name{as_SurvivalProbabilities}
\alias{as_SurvivalProbabilities}
\alias{as_SurvivalProbabilities.ZeroHazardRate}
\title{Coerce to InterestRate}
\usage{
as_SurvivalProbabilities(x, ...)

\method{as_SurvivalProbabilities}{ZeroHazardRate}(x, d1, d2, ...)
}
\arguments{
\item{x}{object to coerce}

\item{...}{other parameters passed to methods}

\item{d1}{a \code{Date} vector containing the as of date}

\item{d2}{a \code{Date} vector containing the date to which the survival probability
applies}
}
\value{
an \code{SurvivalProbabilities} object
}
\description{
You can coerce objects to the \code{SurvivalProbabilities} class using this method.
}
\examples{
curve_specs <- CDSMarkitSpec(
  rating = "AAA",
  region = "Japan",
  sector = "Utilities"
)
HR <- ZeroHazardRate(values = c(0.04, 0.05), compounding = c(2, 4),
day_basis = 'act/365', specs = curve_specs)
as_SurvivalProbabilities(HR, lubridate::ymd(20160202), lubridate::ymd(20160302))
}
