% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{as_tibble.CreditCurve}
\alias{as_tibble.CreditCurve}
\title{CreditCurve attributes as a data frame}
\usage{
\method{as_tibble}{CreditCurve}(x, ...)
}
\arguments{
\item{x}{a \code{CreditCurve} object}

\item{...}{other parameters that are not used by this methods}
}
\value{
a \code{tibble} with two columns named \code{Years} and \code{Zero Hazard Rates}.
}
\description{
Create a \code{tibble} that contains the pillar point maturities in years (using
the \code{act/365} convention) and the corresponding continuously compounded zero
rates.
}
\seealso{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
