% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{as_tibble.VolQuotes}
\alias{as_tibble.VolQuotes}
\title{VolQuotes attributes as a data frame}
\usage{
\method{as_tibble}{VolQuotes}(x, ...)
}
\arguments{
\item{x}{a \code{VolQuotes} object}

\item{...}{other parameters that are not used by this methods}
}
\value{
a \code{tibble} that contains the volatility per maturity and strike.
}
\description{
Create a \code{tibble} that contains the pillar point maturities and strikes for
VolQuotes object with the corresponding volatility.
}
\seealso{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
