% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero-curve-class.R
\name{as_tibble.ZeroCurve}
\alias{as_tibble.ZeroCurve}
\title{ZeroCurve attributes as a data frame}
\usage{
\method{as_tibble}{ZeroCurve}(x, ...)
}
\arguments{
\item{x}{a \code{ZeroCurve} object}

\item{...}{other parameters that are not used by this methods}
}
\value{
a \code{tibble} with two columns named \code{Years} and \code{Zeros}.
}
\description{
Create a \code{tibble} that contains the pillar point maturities in years (using
the \code{act/365} convention) and the corresponding continuously compounded zero
rates.
}
\examples{
library(tibble)
zc <- build_zero_curve()
as_tibble(zc)
}
\seealso{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
