% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-methods.R
\name{as_ZeroHazardCurve}
\alias{as_ZeroHazardCurve}
\title{Bootstraps Hazard Rate from a CDS curve
Using \href{https://www.rdocumentation.org/packages/credule/versions/0.1.3}{credule package.}}
\usage{
as_ZeroHazardCurve(x, ...)
}
\arguments{
\item{cds_curve}{An object of type \code{CDSCurve}}

\item{zero_curve}{An object of type \code{ZeroCurve}}

\item{num_timesteps_pa}{It represents the number of timesteps used to perform the numerical integral
required while computing the default leg value. It is shown that a monthly discretisation
usually gives a good approximation (Ref. Valuation of Credit Default Swaps, Dominic O Kane and
Stuart Turnbull)}

\item{accrued_premium}{If set to TRUE, the accrued premium will be taken into account in the calculation of the premium leg value.}
}
\value{
An object of type \code{HazardCurve}
}
\description{
Bootstraps Hazard Rate from a CDS curve
Using \href{https://www.rdocumentation.org/packages/credule/versions/0.1.3}{credule package.}
}
\examples{
zero_curve <- build_zero_curve()
specs <- CDSMarkitSpec(rating = "AAA", region = "Japan", sector = "Utilities")
cds_curve <- CDSCurve(
 reference_date = zero_curve$reference_date,
   tenors = c(1, 3, 5, 7), spreads = c(0.0050, 0.0070, 0.0090, 0.0110), lgd = .6,
   premium_frequency = 4, specs = curve_specs)
as_ZeroHazardCurve(x = cds_curve, zero_curve = zero_curve)

}
