% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_fx_rates}
\alias{build_fx_rates}
\title{Build \code{FXRates} pricing environment}
\usage{
build_fx_rates()
}
\value{
a \code{FXRates} object using data from \code{fx.csv}
}
\description{
This creates a \code{\link{FXRates}} object from the example data set
\code{zerocurves.csv}.
}
\examples{
build_fx_rates()
}
\seealso{
Other build object helpers: \code{\link{build_pricing_env}},
  \code{\link{build_zero_curves}},
  \code{\link{build_zero_curve}},
  \code{\link{fmdata_example}}
}
