% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_pricing_env}
\alias{build_pricing_env}
\title{Build a pricing environment}
\usage{
build_pricing_env()
}
\value{
a \code{PricingEnv} object using example data
}
\description{
This creates a \code{\link{PricingEnv}} object from the example data sets
\code{zerocurves.csv} and \code{fx.csv}.
}
\examples{
build_pricing_env()
}
\seealso{
Other build object helpers: \code{\link{build_fx_rates}},
  \code{\link{build_zero_curves}},
  \code{\link{build_zero_curve}},
  \code{\link{fmdata_example}}
}
