% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_vol_quotes}
\alias{build_vol_quotes}
\title{Build a \code{VolQuotes} object from an example data set}
\usage{
build_vol_quotes()
}
\value{
a \code{VolQuotes} object from package built-in data
}
\description{
This creates an object of class \code{VolQuotes} from the example data set
\code{volsurface.csv}.
}
\examples{
build_vol_quotes()
}
\seealso{
Other build object helpers: \code{\link{build_vol_surface}},
  \code{\link{build_zero_curve}}
}
\concept{build object helpers}
