% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_vol_surface}
\alias{build_vol_surface}
\title{Build a \code{VolSurface} from an example date set}
\usage{
build_vol_surface()
}
\value{
a \code{VolSurface} object using data from \code{volsurface.csv}
}
\description{
This creates a \code{\link{VolSurface}} object from the example data set
\code{volsurface.csv}.
}
\examples{
build_vol_surface()
}
\seealso{
Other build vol object helpers: \code{\link{build_vol_quotes}}
}
\concept{build vol object helpers}
