% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_zero_curve}
\alias{build_zero_curve}
\title{Build a \code{ZeroCurve} from example data set}
\usage{
build_zero_curve(interpolation = NULL)
}
\arguments{
\item{interpolation}{an \code{Interpolation} object}
}
\value{
a \code{ZeroCurve} object using data from \code{zerocurve.csv}
}
\description{
This creates a \code{\link{ZeroCurve}} object from the example data set
\code{zerocurve.csv}.
}
\examples{
build_zero_curve(LogDFInterpolation())
}
