% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_zero_curves}
\alias{build_zero_curves}
\title{Build \code{ZeroCurves} pricing environment}
\usage{
build_zero_curves(interpolation = NULL)
}
\arguments{
\item{interpolation}{an \code{Interpolation} object}
}
\value{
a \code{ZeroCurves} object using data from \code{zerocurves.csv}
}
\description{
This creates a \code{ZeroCurves} object from the example data set \code{zerocurves.csv}.
}
\examples{
build_zero_curves(LogDFInterpolation())
}
\seealso{
\code{\link[=fmdata_example]{fmdata_example()}}
}
