% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash-flow-class.R
\name{CashFlow}
\alias{CashFlow}
\title{Create a CashFlow}
\usage{
CashFlow(dates, monies)
}
\arguments{
\item{dates}{a \link[=Date]{Date} vector with either the same length as \code{monies}
or a vector of length one that is recycled}

\item{monies}{a \link[=MultiCurrencyMoney]{MultiCurrencyMoney} object}
}
\value{
a \code{CashFlow} object that extends \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This allows you to create a \code{CashFlow} object.
}
\examples{
CashFlow(as.Date("2017-11-15"),
  MultiCurrencyMoney(list(SingleCurrencyMoney(1, AUD())))
)
}
\seealso{
Other money functions: \code{\link{MultiCurrencyMoney}},
  \code{\link{SingleCurrencyMoney}},
  \code{\link{is.CashFlow}},
  \code{\link{is.MultiCurrencyMoney}},
  \code{\link{is.SingleCurrencyMoney}}
}
