% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{CDSCurve}
\alias{CDSCurve}
\title{Builds a \code{CDSCurve}}
\usage{
CDSCurve(reference_date, tenors, spreads, lgd, premium_frequency, specs)
}
\arguments{
\item{reference_date}{the curve's reference date as a \link[base:Date]{base::Date}}

\item{tenors}{a numeric vector of pillar points time steps expressed in years}

\item{spreads}{a numeric vector of creadit default spreads expressed in
decimals. Must be the same length as \code{tenors}}

\item{lgd}{the loss given default associated with the curve as supplied by
Markit and expressed as a decimal value}

\item{premium_frequency}{represents the number of premiums payments per annum
expressed as an integer. Must be one of 1, 2, 4 or 12.}

\item{specs}{CDS curve specifications that inherits from \code{\link[=CDSSpec]{CDSSpec()}}}
}
\value{
An object of type \code{CDSCurve}
}
\description{
This will allow you to create an instance of a CDS curve.
}
\examples{
curve_specs <- CDSMarkitSpec(
  rating = "AAA",
  region = "Japan",
  sector = "Utilities"
)

CDSCurve(
  as.Date("2019-06-29"),
  tenors = c(1, 3, 5, 7),
  spreads = c(0.0050, 0.0070, 0.0090, 0.0110),
  lgd = 0.6,
  premium_frequency = 4,
  specs = curve_specs
)
}
\seealso{
Other CDS curve helpers: \code{\link{CDSMarkitSpec}},
  \code{\link{CDSSingleNameSpec}}, \code{\link{CDSSpec}},
  \code{\link{HazardCurve}}, \code{\link{SurvivalCurve}},
  \code{\link{is.CDSCurve}}, \code{\link{is.CDSSpec}}
}
\concept{CDS curve helpers}
