% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{CDSMarkitSpec}
\alias{CDSMarkitSpec}
\title{Build a \code{CDSMarkitSpec}}
\usage{
CDSMarkitSpec(rating, region, sector)
}
\arguments{
\item{rating}{valid options are "AAA", "AA", "A", "BBB", "BB", "B", "CCC"}

\item{region}{valid options are "AsiaExJapan", "EastEurope", "Europe",
"Japan", "LatinAmerica", "NorthAmerica", "MiddleEast", "Oceania"}

\item{sector}{valid options are "BasicMaterials", "ConsumerGoods",
"ConsumerServices", "Energy", "Financials", "Government", "Healtcare",
"Technology", "TeleCom", "Utilities"}
}
\value{
An object of type \code{CDSMarkitSpec}
}
\description{
A subclass of \code{\link[=CDSSpec]{CDSSpec()}}, only for Markit sector curves. Note that the
paramter \code{rank} is fixed to be "SNR", as per Markit's methodology documents
}
\examples{
CDSMarkitSpec(rating = "AAA", region = "Japan", sector = "Utilities")
}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSSingleNameSpec}}, \code{\link{CDSSpec}},
  \code{\link{SurvivalProbabilities}},
  \code{\link{ZeroHazardRate}}, \code{\link{is.CDSCurve}},
  \code{\link{is.CDSSpec}}
}
\concept{CDS curve helpers}
