% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{CDSSingleNameSpec}
\alias{CDSSingleNameSpec}
\title{Builds a \code{CDSSingleNameSpec}}
\usage{
CDSSingleNameSpec(rank, name)
}
\arguments{
\item{rank}{Seniority of the reference debt. Must be one of the following
options: "SNR" for Senior, "SubTier3" for Subordinate Tier 3,
"SubUpperTier2" for Subordinate Upper Tier 2, "SubLowerTier2" for
Subordinate Lower Tier 2 "SubTier1" for Subordinate Tier 1}

\item{name}{Reference debt issuer. Must be a string.}
}
\value{
An object of type \code{CDSSingleNameSpec}
}
\description{
A subclass of \code{\link[=CDSSpec]{CDSSpec()}}, that implements specifications for single name
CDS curves
}
\examples{
CDSSingleNameSpec(rank = "SNR", name = "Westpac")
}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSMarkitSpec}}, \code{\link{CDSSpec}},
  \code{\link{HazardCurve}}, \code{\link{SurvivalCurve}}
}
\concept{CDS curve helpers}
