% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{CDSSingleNameSpecs}
\alias{CDSSingleNameSpecs}
\title{Builds a \code{CDSSingleNameSpecs}}
\usage{
CDSSingleNameSpecs(rank, name)
}
\arguments{
\item{rank}{Seniority of the reference debt. Must be one of the following
options: "SNR" for Senior, "SubTier3" for Subordinate Tier 3,
"SubUpperTier2" for Subordinate Upper Tier 2, "SubLowerTier2" for
Subordinate Lower Tier 2 "SubTier1" for Subordinate Tier 1}

\item{name}{Reference debt issuer. Must be a string.}
}
\value{
An object of type \code{CDSSingleNameSpecs}
}
\description{
A subclass of \code{\link[=CDSSpecs]{CDSSpecs()}}, that implements specifications for single name
CDS curves
}
\examples{
CDSSingleNameSpecs(rank = "SNR", name = "Westpac")
}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSMarkitSpecs}}, \code{\link{CDSSpecs}},
  \code{\link{HazardCurve}}, \code{\link{SurvivalCurve}}
}
\concept{CDS curve helpers}
