% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{CDSSpecs}
\alias{CDSSpecs}
\title{Build a \code{CDSSpecs}}
\usage{
CDSSpecs(rank, ..., subclass = NULL)
}
\arguments{
\item{rank}{Seniority of the reference debt. Must be one of the following
options: "SNR" for Senior, "SubTier3" for Subordinate Tier 3,
"SubUpperTier2" for Subordinate Upper Tier 2, "SubLowerTier2" for
Subordinate Lower Tier 2 "SubTier1" for Subordinate Tier 1}

\item{...}{parameters passed to other \code{CDSSpecs} constructors}

\item{subclass}{the name of a \code{CDSSpecs} subclass. Defaults to \code{NULL}}
}
\value{
Object of type \code{CDSSpecs}
}
\description{
This class will enable you to specify CDS curves. It is used by
\code{\link[=SurvivalCurve]{SurvivalCurve()}} and \code{\link[=HazardCurve]{HazardCurve()}}.
}
\examples{
CDSSpecs(rank = "SubTier3")
}
