% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{CreditCurve}
\alias{CreditCurve}
\title{CreditCurve class}
\usage{
CreditCurve(survival_probabilities, reference_date, interpolation, specs)
}
\arguments{
\item{survival_probabilities}{a \code{\link{SurvivalProbabilities}} object. These are converted to
continuously compounded zero coupon interest rates with an \code{act/365} day
basis for internal storage purposes}

\item{reference_date}{a \code{Date} object}

\item{interpolation}{an \code{\link{Interpolation}} object}

\item{specs}{CDS curve specifications that inherits from \code{\link[=CDSSpec]{CDSSpec()}}}
}
\value{
a \code{CreditCurve} object
}
\description{
A class that defines the bare bones of a credit curve pricing
structure.
}
\details{
A term structure of credit spread is a curve showing
several credit spreads across different contract lengths (2 month,
2 year, 20 year, etc...) for a similar debt contract. The curve shows the
relation between the (level of) crdit spread and the
time to maturity, known as the "term", of the debt for a given borrower in a
given currency. When the effect of coupons on spreads are stripped away, one has a
zero-coupon credit curve.

The following interpolation schemes are supported by \code{ZeroCurve}:
\code{ConstantInterpolation}, \code{LinearInterpolation}, \code{LogDFInterpolation} and
\code{CubicInterpolation}. Points outside the calibration region use constant
extrapolation on the zero hazard rate.
}
\examples{
curve_specs <- CDSMarkitSpec(rating = "AAA", region = "Japan", sector = "Utilities")
zero_curve <- build_zero_curve()
ref_date <- zero_curve$reference_date
specs <- CDSMarkitSpec(rating = "AAA", region = "Japan", sector = "Utilities")
cds_curve <- CDSCurve(reference_date = ref_date, tenors = c(1, 3, 5, 7),
spreads = c(0.0050, 0.0070, 0.0090, 0.0110), lgd = .6, premium_frequency = 4,
specs = curve_specs)
sp <- as_SurvivalProbabilities(x = cds_curve, zero_curve = zero_curve)
CreditCurve(survival_probabilities = sp, reference_date =ref_date,
 interpolation =  CubicInterpolation(), specs = curve_specs)
}
\seealso{
\link{Interpolation}
}
