% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-class.R, R/currency-pair-class.R
\name{CurrencyConstructors}
\alias{CurrencyConstructors}
\alias{AUD}
\alias{EUR}
\alias{GBP}
\alias{JPY}
\alias{NZD}
\alias{USD}
\alias{CHF}
\alias{HKD}
\alias{NOK}
\alias{CurrencyConstructors}
\alias{AUDUSD}
\alias{EURUSD}
\alias{NZDUSD}
\alias{GBPUSD}
\alias{USDJPY}
\alias{GBPJPY}
\alias{EURGBP}
\alias{AUDNZD}
\alias{EURCHF}
\alias{USDCHF}
\alias{USDHKD}
\alias{EURNOK}
\alias{USDNOK}
\title{Handy Currency constructors}
\usage{
AUD()

EUR()

GBP()

JPY()

NZD()

USD()

CHF()

HKD()

NOK()

AUDUSD()

EURUSD()

NZDUSD()

GBPUSD()

USDJPY()

GBPJPY()

EURGBP()

AUDNZD()

EURCHF()

USDCHF()

USDHKD()

EURNOK()

USDNOK()
}
\description{
These constructors use the following conventions:

These handy \code{CurrencyPair} constructors use their \link[=CurrencyConstructors]{single currency
counterparts} in the obvious fashion.
}
\details{
\tabular{ll}{
\bold{Creator} \tab \bold{Joint calendars} \cr
\code{AUD()} \tab \code{AUSYCalendar} \cr
\code{EUR()} \tab \code{EUTACalendar} \cr
\code{GBP()} \tab \code{GBLOCalendar} \cr
\code{JPY()} \tab \code{JPTOCalendar} \cr
\code{NZD()} \tab \code{NZAUCalendar}, \code{NZWECalendar} \cr
\code{USD()} \tab \code{USNYCalendar} \cr
\code{CHF()} \tab \code{CHZHCalendar} \cr
\code{HKD()} \tab \code{HKHKCalendar} \cr
\code{NOK()} \tab \code{NOOSCalendar}
}
}
\examples{
AUD()
AUDUSD()
}
