% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-pair-class.R
\name{CurrencyPair}
\alias{CurrencyPair}
\title{CurrencyPair class}
\usage{
CurrencyPair(base_ccy, quote_ccy, calendar = NULL)
}
\arguments{
\item{base_ccy}{a \link[=Currency]{Currency} object}

\item{quote_ccy}{a \link[=Currency]{Currency} object}

\item{calendar}{a \link[=JointCalendar]{JointCalendar} object. Defaults to
\code{NULL} which sets this to the joint calendar of the two currencies and
removes any \link[=USNYCalendar]{USNYCalendar} object to allow currency pair
methods to work correctly}
}
\value{
a \code{CurrencyPair} object
}
\description{
Create an object of class \code{CurrencyPair}
}
\examples{
CurrencyPair(AUD(), USD())
}
