% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-pair-class.R
\name{CurrencyPairConstructors}
\alias{CurrencyPairConstructors}
\alias{AUDUSD}
\alias{EURUSD}
\alias{NZDUSD}
\alias{GBPUSD}
\alias{USDJPY}
\alias{GBPJPY}
\alias{EURGBP}
\alias{AUDNZD}
\alias{EURCHF}
\alias{USDCHF}
\alias{USDHKD}
\alias{EURNOK}
\alias{USDNOK}
\title{Handy CurrencyPair constructors}
\usage{
AUDUSD()

EURUSD()

NZDUSD()

GBPUSD()

USDJPY()

GBPJPY()

EURGBP()

AUDNZD()

EURCHF()

USDCHF()

USDHKD()

EURNOK()

USDNOK()
}
\description{
These handy \code{CurrencyPair} constructors use their \link[=CurrencyConstructors]{single currency
counterparts} in the obvious fashion.
}
\examples{
AUDUSD()
}
\seealso{
Other constructors: \code{\link{CurrencyConstructors}},
  \code{\link{iborindices}}, \code{\link{oniaindices}}
}
