% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest-rate-ops.R
\name{DiscountFactor-operators}
\alias{DiscountFactor-operators}
\title{\code{DiscountFactor} operations}
\description{
As the \code{\link{DiscountFactor}} is built with the vctrs package all base
vector operations are available (i.e. \code{c}, \code{[}, ...).
}
\details{
Additionally \code{DiscountFactor} objects may be multiplied and divided
under the following constraints:
\itemize{
\item \code{*}: multiplication of \code{DiscountFactor} objects. The end
date of the first discount factor object must be equivalent to the start
date of the second (or vice versa). Arguments are recycled as necessary.
\item \code{/}: division of \code{DiscountFactor} objects. The start date
date of both arguments must be the same. Arguments are recycled as necessary.
}
}
