% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmbasics-package.r
\docType{package}
\name{fmbasics}
\alias{fmbasics}
\alias{fmbasics-package}
\title{fmbasics: Financial Market Building Blocks}
\description{
Implements basic financial market objects like currencies, currency pairs,
interest rates and interest rate indices. You will be able to use Benchmark
instances of these objects which have been defined using their most common
conventions or those defined by International Swap Dealer Association legal
documentation.
}
