% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-env-class.R
\name{FXRates}
\alias{FXRates}
\title{Create a \code{FXRates} pricing object}
\usage{
FXRates(isos, rates)
}
\arguments{
\item{isos}{a character vector of ISO codes representing currency pairs (e.g.
"AUDUSD") and must not contain duplicate values}

\item{rates}{a numeric vector representing the values of the FX rates. This
must be the same length as \code{isos}}
}
\value{
a \code{FXRates} object that extends a tibble
}
\description{
A \code{FXRates} object is effectively a \code{\link[tibble:tibble]{tibble::tibble()}} containing currency
pair ISOs and their associated values (usually spot FX).
}
\examples{
build_fx_rates()
}
\seealso{
Other pricing environments: \code{\link{PricingEnv}},
  \code{\link{ZeroCurves}}, \code{\link{is.FXRates}},
  \code{\link{is.PricingEnv}}, \code{\link{is.ZeroCurves}},
  \code{\link{pick}}
}
