% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{HazardCurve}
\alias{HazardCurve}
\title{Builds a \code{HazardCurve}}
\usage{
HazardCurve(reference_date, tenors, hazard_rates, specs)
}
\arguments{
\item{reference_date}{the curve's reference date as a \link[base:Date]{base::Date}}

\item{tenors}{a numeric vector of pillar points time steps expressed in years}

\item{hazard_rates}{a vector of hazard rates corresponding to each time step
in \code{tenors}}

\item{specs}{CDS curve specifications that inherits from \code{\link[=CDSSpec]{CDSSpec()}}}
}
\value{
returns an object of type \code{hazard_rates}
}
\description{
This will allow you to create a harzard rate curve. This will typically be
bootstrapped or implied from a \code{\link[=CDSCurve]{CDSCurve()}} or \code{\link[=SurvivalCurve]{SurvivalCurve()}}.
}
\examples{
curve_specs <- CDSMarkitSpec(
  rating = "AAA",
  region = "Japan",
  sector = "Utilities"
)

HazardCurve(
  reference_date = as.Date("2019-06-29"),
  tenors = c(1, 3, 5, 7),
  hazard_rates = c(0.05, 0.05, 0.05, 0.05),
  specs = curve_specs
)
}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSMarkitSpec}},
  \code{\link{CDSSingleNameSpec}}, \code{\link{CDSSpec}},
  \code{\link{SurvivalCurve}}, \code{\link{is.CDSCurve}},
  \code{\link{is.CDSSpec}}
}
\concept{CDS curve helpers}
