% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices-class.R
\name{indexcheckers}
\alias{indexcheckers}
\alias{is.Index}
\alias{is.IborIndex}
\alias{is.CashIndex}
\title{Index class checkers}
\usage{
is.Index(x)

is.IborIndex(x)

is.CashIndex(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{TRUE} if object inherits from tested class
}
\description{
Index class checkers
}
\examples{
is.Index(AONIA())
is.CashIndex(AONIA())
is.IborIndex(AONIA())
}
