% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest-rate-ops.R
\name{InterestRate-operators}
\alias{InterestRate-operators}
\title{\code{InterestRate} operations}
\description{
The \code{\link{InterestRate}} class is built using the vctrs package.
Therefore all base vector operations can be performed on or with
\code{\link{InterestRate}} objects.
}
\details{
Some care has to be taken when using arithmetic and comparison operators:
\itemize{
\item Arithmetic operations on \code{InterestRate} objects: Where both
arguments are \code{InterestRate} objects,their compounding and day
basis frequency must match.
Operations between numeric values and an \code{InterestRate} object are
performed directly on the \code{rate} field.
Arguments are recycled as necessary.
\item \code{<, >, <=, >=, ==, !=}: To enable comparison both vectors are
converted to infinite compounding and act/365 day basis. The
resulting \code{rate} field is used for comparison.
}
}
