% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{interpolate.CreditCurve}
\alias{interpolate.CreditCurve}
\title{Interpolate a \code{CreditCurve}}
\usage{
\method{interpolate}{CreditCurve}(x, at, ...)
}
\arguments{
\item{x}{a \code{CreditCurve} object}

\item{at}{a non-negative numeric vector representing the years at which to
interpolate the Credit curve}

\item{...}{unused in this method}
}
\value{
a numeric vector of zero rates (continuously compounded, act/365)
}
\description{
There are two key interpolation schemes available in the \code{stats} package:
constant and linear interpolation via \code{\link[stats:approxfun]{stats::approxfun()}} and
spline interpolation via \code{\link[stats:splinefun]{stats::splinefun()}}. The \code{interpolate()} method
is a simple wrapper around these methods that are useful for the purposes
of interpolation financial market objects like credit curves.
}
\examples{
zc <- build_zero_curve(LogDFInterpolation())
interpolate(zc, c(1.5, 3))
}
\seealso{
Other interpolate functions: \code{\link{interpolate.VolSurface}},
  \code{\link{interpolate.ZeroCurve}},
  \code{\link{interpolate_dfs.CreditCurve}},
  \code{\link{interpolate_zeros.CreditCurve}},
  \code{\link{interpolate}}
}
\concept{interpolate functions}
