% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero-curve.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate values from an object}
\usage{
interpolate(x, ...)
}
\arguments{
\item{x}{the object to interpolate.}

\item{...}{other parameters that defines how to interpolate the object}
}
\value{
an interpolated value or set of values
}
\description{
Interpolate values from an object
}
