% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{interpolate.VolSurface}
\alias{interpolate.VolSurface}
\title{Interpolate a \code{VolSurface} object.}
\usage{
\method{interpolate}{VolSurface}(x, at, ...)
}
\arguments{
\item{x}{object of class \code{VolSurface} to be interpolated.}

\item{at}{indicates the coordinates at which the interpolation is performed.
\code{at} should be given as a named list of length two, where the two members are
vectors with the same length and are named \code{term} and \code{strike}. e.g.
list(term = c(1, 2), strike = c(72, 92)).}

\item{...}{unused in this model.}
}
\value{
\code{numeric} vector with length equal to the length of \code{at} members.
}
\description{
This method is used to interpolate a \code{VolSurface} object at multiple points of
the plane. The interpolation depends on the type of the surface, if the vols are
given by strikes, delta, moneyness.
}
\examples{
vol_surface <- fmbasics::build_vol_surface()
interpolation_points <- list(term = c(as.Date("2020-03-31"), as.Date("2021-03-31")),
strike = c(40, 80))
implied_vols <- fmbasics::interpolate(x = vol_surface, at  = interpolation_points)
}
\seealso{
Other interpolate functions: \code{\link{interpolate.ZeroCurve}},
  \code{\link{interpolate_dfs}},
  \code{\link{interpolate_zeros}},
  \code{\link{interpolate}}
}
\concept{interpolate functions}
