% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-class.R
\name{Interpolation}
\alias{Interpolation}
\alias{ConstantInterpolation}
\alias{LogDFInterpolation}
\alias{LinearInterpolation}
\alias{CubicInterpolation}
\alias{LinearCubicTimeVarInterpolation}
\title{Interpolation}
\usage{
ConstantInterpolation()

LogDFInterpolation()

LinearInterpolation()

CubicInterpolation()

LinearCubicTimeVarInterpolation()
}
\value{
an object that inherits from the \code{Interpolation} class.
}
\description{
These are lightweight interpolation classes that are used to specify
typical financial market interpolation schemes. Their behaviour is
dictated by the object in which they defined.
}
\examples{
ConstantInterpolation()
}
