% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash-flow-class.R
\name{is.CashFlow}
\alias{is.CashFlow}
\title{Inherits from CashFlow}
\usage{
is.CashFlow(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{CashFlow} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{CashFlow} class
}
\examples{
is.CashFlow(CashFlow(as.Date("2017-11-15"),
  MultiCurrencyMoney(list(SingleCurrencyMoney(1, AUD())))))
}
\seealso{
Other money functions: \code{\link{CashFlow}},
  \code{\link{MultiCurrencyMoney}},
  \code{\link{SingleCurrencyMoney}},
  \code{\link{is.MultiCurrencyMoney}},
  \code{\link{is.SingleCurrencyMoney}}
}
