% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{is.CDSCurve}
\alias{is.CDSCurve}
\title{Inherits from CDSCurve}
\usage{
is.CDSCurve(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{CDSCurve} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{CDSCurve} class
}
\examples{
curve_specs <- CDSMarkitSpec(
  rating = "AAA",
  region = "Japan",
  sector = "Utilities"
)
cds_curve <- CDSCurve(
  as.Date("2019-06-29"),
  tenors = c(1, 3, 5, 7),
  spreads = c(0.0050, 0.0070, 0.0090, 0.0110),
  lgd = 0.6,
  premium_frequency = 4,
  specs = curve_specs
)
is.CDSCurve(cds_curve)
}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSMarkitSpec}},
  \code{\link{CDSSingleNameSpec}}, \code{\link{CDSSpec}},
  \code{\link{SurvivalProbabilities}},
  \code{\link{ZeroHazardRate}}, \code{\link{is.CDSSpec}}
}
\concept{CDS curve helpers}
