% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{is.CDSSpec}
\alias{is.CDSSpec}
\title{Inherits from CDSSpec}
\usage{
is.CDSSpec(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{CDSSpec} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{CDSSpec} class
}
\examples{
curve_specs <- CDSMarkitSpec(
  rating = "AAA",
  region = "Japan",
  sector = "Utilities"
)
is.CDSSpec(curve_specs)
}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSMarkitSpec}},
  \code{\link{CDSSingleNameSpec}}, \code{\link{CDSSpec}},
  \code{\link{SurvivalProbabilities}},
  \code{\link{ZeroHazardRate}}, \code{\link{is.CDSCurve}}
}
\concept{CDS curve helpers}
