% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{is.CreditCurve}
\alias{is.CreditCurve}
\title{Inherits from CreditCurve}
\usage{
is.CreditCurve(x)
}
\arguments{
\item{x}{an R Object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{CreditCurve} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{CreditCurve} class
}
