% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-class.R
\name{is.Currency}
\alias{is.Currency}
\title{Inherits from Currency}
\usage{
is.Currency(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{Currency} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{Currency} class
}
\examples{
is.Currency(AUD())
}
