% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-pair-class.R
\name{is.CurrencyPair}
\alias{is.CurrencyPair}
\title{Inherits from \code{CurrencyPair} class}
\usage{
is.CurrencyPair(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{CurrencyPair} class; otherwise \code{FALSE}
}
\description{
Inherits from \code{CurrencyPair} class
}
\examples{
is.CurrencyPair(AUDUSD())
}
