% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest-rate-classes.R
\name{is.DiscountFactor}
\alias{is.DiscountFactor}
\title{Inherits from DiscountFactor}
\usage{
is.DiscountFactor(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{DiscountFactor} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{DiscountFactor} class
}
\examples{
is.DiscountFactor(DiscountFactor(0.97, Sys.Date(), Sys.Date() + 30))
}
