% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest-rate-classes.R
\name{is.InterestRate}
\alias{is.InterestRate}
\title{Inherits from InterestRate}
\usage{
is.InterestRate(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{InterestRate} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{InterestRate} class
}
\examples{
is.InterestRate(InterestRate(0.04, 2, "act/365"))
}
