% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero-curve-class.R
\name{is.Interpolation}
\alias{is.Interpolation}
\alias{is.ConstantInterpolation}
\alias{is.LogDFInterpolation}
\alias{is.LinearInterpolation}
\alias{is.CubicInterpolation}
\title{Check Interpolation class}
\usage{
is.Interpolation(x)

is.ConstantInterpolation(x)

is.LogDFInterpolation(x)

is.LinearInterpolation(x)

is.CubicInterpolation(x)
}
\arguments{
\item{x}{an object}
}
\value{
a logical flag
}
\description{
These methods check whether an interpolation is of a particular scheme.
}
\examples{
is.Interpolation(CubicInterpolation())
is.CubicInterpolation(CubicInterpolation())
}
