% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/money-class.R
\name{is.MultiCurrencyMoney}
\alias{is.MultiCurrencyMoney}
\title{Inherits from MultiCurrencyMoney}
\usage{
is.MultiCurrencyMoney(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{MultiCurrencyMoney} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{MultiCurrencyMoney} class
}
\examples{
is.MultiCurrencyMoney(MultiCurrencyMoney(list(SingleCurrencyMoney(1, AUD()))))
}
\seealso{
Other money functions: \code{\link{CashFlow}},
  \code{\link{MultiCurrencyMoney}},
  \code{\link{SingleCurrencyMoney}},
  \code{\link{is.CashFlow}},
  \code{\link{is.SingleCurrencyMoney}}
}
