% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-env-class.R
\name{is.PricingEnv}
\alias{is.PricingEnv}
\title{Inherits from PricingEnv}
\usage{
is.PricingEnv(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{PricingEnv} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{PricingEnv} class
}
\examples{
is.PricingEnv(build_pricing_env())
}
\seealso{
Other pricing environments: \code{\link{FXRates}},
  \code{\link{PricingEnv}}, \code{\link{ZeroCurves}},
  \code{\link{is.FXRates}}, \code{\link{is.ZeroCurves}},
  \code{\link{pick}}
}
