% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/money-class.R
\name{is.SingleCurrencyMoney}
\alias{is.SingleCurrencyMoney}
\title{Inherits from SingleCurrencyMoney}
\usage{
is.SingleCurrencyMoney(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{SingleCurrencyMoney} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{SingleCurrencyMoney} class
}
\examples{
is.SingleCurrencyMoney(SingleCurrencyMoney(1:5, AUD()))
}
\seealso{
Other money functions: \code{\link{CashFlow}},
  \code{\link{MultiCurrencyMoney}},
  \code{\link{SingleCurrencyMoney}},
  \code{\link{is.CashFlow}},
  \code{\link{is.MultiCurrencyMoney}}
}
