% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{is.SurvivalProbabilities}
\alias{is.SurvivalProbabilities}
\title{Inherits from SurvivalProbabilities}
\usage{
is.SurvivalProbabilities(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{SurvivalProbabilities} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{SurvivalProbabilities} class
}
\examples{
is.SurvivalProbabilities(SurvivalProbabilities(0.97, Sys.Date(), Sys.Date() + 30, CDSSpec("Empty")))
}
