% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{is.VolQuotes}
\alias{is.VolQuotes}
\title{Inherits from VolQuotes}
\usage{
is.VolQuotes(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{VolQuotes} class; otherwise \code{FALSE}
}
\description{
Checks whether the object inherits from \code{VolQuotes} class
}
