% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{is.VolSurface}
\alias{is.VolSurface}
\title{Inherits from VolSurface}
\usage{
is.VolSurface(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{VolSurface} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{VolSurface} class
}
