% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero-curve-class.R
\name{is.ZeroCurve}
\alias{is.ZeroCurve}
\title{Inherits from ZeroCurve}
\usage{
is.ZeroCurve(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{ZeroCurve} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{ZeroCurve} class
}
\examples{
zc_df <- fmdata_example("zerocurve.csv")
values <- zc_df[["dfs"]]
starts <- as.Date(as.character(zc_df[["start"]]), "\%Y\%m\%d")
ends <- as.Date(as.character(zc_df[["end"]]), "\%Y\%m\%d")
dfs <- DiscountFactor(values, starts, ends)
zc <- ZeroCurve(dfs, starts[1], LogDFInterpolation())
is.ZeroCurve(zc)
}
