% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-env-class.R
\name{is.ZeroCurves}
\alias{is.ZeroCurves}
\title{Inherits from ZeroCurves}
\usage{
is.ZeroCurves(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{ZeroCurves} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{ZeroCurves} class
}
\examples{
is.ZeroCurves(build_zero_curves())
}
\seealso{
Other pricing environments: \code{\link{FXRates}},
  \code{\link{PricingEnv}}, \code{\link{ZeroCurves}},
  \code{\link{is.FXRates}}, \code{\link{is.PricingEnv}},
  \code{\link{pick}}
}
