% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{is.ZeroHazardRate}
\alias{is.ZeroHazardRate}
\title{Inherits from ZeroHazardRate}
\usage{
is.ZeroHazardRate(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{ZeroHazardRate} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{ZeroHazardRate} class
}
\examples{
is.ZeroHazardRate(ZeroHazardRate(0.04, 2, "act/365", CDSSpec("Empty")))
}
