% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-class.R
\name{iso}
\alias{iso}
\title{ISO}
\usage{
iso(x)
}
\arguments{
\item{x}{object from which to extract an ISO}
}
\value{
a string of the ISO
}
\description{
Get ISO from an object. The default method assumes the ISO can be accessed
as if it were an attribute with name \code{iso} (e.g. \code{x$iso}).
}
\examples{
iso(AUD())
}
