% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-class.R
\name{linear_cubic_interp}
\alias{linear_cubic_interp}
\title{Linear-Cubic Interpolation}
\usage{
linear_cubic_interp(interp_data, x0, y0)
}
\arguments{
\item{interp_data}{\code{data.frame} object with three vectors x, y and z}

\item{x0}{numeric vector containing the points at which to perform the interpolation along the first dimension}

\item{y0}{numeric vector containing the points at which to perform the interpolation along the second dimension}
}
\description{
This function performs a two-dimentional interpolation linear on the first dimension
and natural cubic spline on the second dimension.
}
\keyword{internal}
