% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/money-class.R
\name{MultiCurrencyMoney}
\alias{MultiCurrencyMoney}
\title{MultiCurrencyMoney}
\usage{
MultiCurrencyMoney(values, currencies)
}
\arguments{
\item{values}{a vector of numeric values}

\item{currencies}{a list of \link[=Currency]{Currency} objects with the same
length as \code{values} or a single object that is recycled}
}
\value{
a \code{MultiCurrencyMoney} object that extends \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This class associated a vector of numeric values with a list of currencies.
This can be useful for example to store value of cash flows.
}
\examples{
MultiCurrencyMoney(c(100, 200), list(AUD(), USD()))
}
\seealso{
Other money functions: \code{\link{CashFlow}},
  \code{\link{SingleCurrencyMoney}},
  \code{\link{is.CashFlow}},
  \code{\link{is.MultiCurrencyMoney}},
  \code{\link{is.SingleCurrencyMoney}}
}
