% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-env-class.R
\name{pick}
\alias{pick}
\title{Pick out element from PricingEnv}
\usage{
pick(from, what, ...)
}
\arguments{
\item{from}{a \link{PricingEnv}}

\item{what}{a string specifying where the pricing element can be found. This
corresponds to the \code{PricingEnv} field names which are currently
\code{zero_curves} and \code{fx_rates}.}

\item{...}{other parameters to full specify the pricing element to extract.
For \code{zero_curves} this includes the parameters \code{currency} and \code{index} which
when concatenated with an underscore separator specifies the name of the
\code{ZeroCurve} to extract. For \code{fx_rates} this includes \code{iso} which specifies
the ISO code of the currency pair of the FX rate (numeric) that is to be
extracted.}
}
\value{
a \code{ZeroCurve} when \code{what} is "zero_curves" and a number when \code{what}
is "fx_rates".
}
\description{
This will allow you to pick an element out of a \link{PricingEnv} given
a set of parameters.
}
\examples{
pe <- build_pricing_env()
pick(pe, "zero_curves", "CC1", "3M")
pick(pe, "fx_rates", "AUDUSD")
}
\seealso{
Other pricing environments: \code{\link{FXRates}},
  \code{\link{PricingEnv}}, \code{\link{ZeroCurves}},
  \code{\link{is.FXRates}}, \code{\link{is.PricingEnv}},
  \code{\link{is.ZeroCurves}}
}
