% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-env-class.R
\name{PricingEnv}
\alias{PricingEnv}
\title{Create a \code{PricingEnv}}
\usage{
PricingEnv(zero_curves, fx_rates, reference_date = NULL)
}
\arguments{
\item{zero_curves}{a \code{\link{ZeroCurves}} object}

\item{fx_rates}{a \code{\link{FXRates}} object}

\item{reference_date}{the \code{Date} on which these pricing objects are snapped.
Defaults to \code{NULL} in which case this evaluates to the \code{reference_date} of
the first zero curve in \code{zero_curves}.}
}
\value{
a \code{PricingEnv} object
}
\description{
A \code{PricingEnv} is a container of objects that are used for the pricing and
valuation of financial market contracts. Supported pricing objects include:
\itemize{
\item \code{\link{ZeroCurves}} and
\item \code{\link{FXRates}}
}
}
\examples{
build_pricing_env()
}
