% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/money-class.R
\name{SingleCurrencyMoney}
\alias{SingleCurrencyMoney}
\title{SingleCurrencyMoney}
\usage{
SingleCurrencyMoney(value, currency)
}
\arguments{
\item{value}{a numeric vector of values}

\item{currency}{a single \link[=Currency]{Currency} object}
}
\value{
a \code{SingleCurrencyMoney} object
}
\description{
This class associates a numeric vector with a currency.
}
\examples{
SingleCurrencyMoney(1:5, AUD())
}
\seealso{
Other money functions: \code{\link{MultiCurrencyMoney}},
  \code{\link{is.MultiCurrencyMoney}},
  \code{\link{is.SingleCurrencyMoney}}
}
