% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{SurvivalCurve}
\alias{SurvivalCurve}
\title{Builds a \code{SurvivalCurve}}
\usage{
SurvivalCurve(reference_date, tenors, probabilities, specs)
}
\arguments{
\item{reference_date}{the curve's reference date as a \link[base:Date]{base::Date}}

\item{tenors}{a numeric vector of pillar points time steps expressed in years}

\item{probabilities}{a vector of survival probabilities corresponding to each
time step in \code{tenors}.}

\item{specs}{CDS curve specifications that inherits from \code{\link[=CDSSpecs]{CDSSpecs()}}}
}
\value{
returns an object of type \code{SurvivalCurve}
}
\description{
This will allow you to create a survival probability curve. This will
typically be bootstrapped from a \code{\link[=CDSCurve]{CDSCurve()}}.
}
\examples{
curve_specs <- CDSMarkitSpecs(
  rating = "AAA",
  region = "Japan",
  sector = "Utilities"
)

SurvivalCurve(
  reference_date = as.Date("2019-06-29"),
  tenors = c(1, 3, 5, 7),
  probabilities = c(0.99, 0.99, 0.99, 0.99),
  specs = curve_specs
)
}
