% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedit-ops.R
\name{SurvivalProbabilities-operators}
\alias{SurvivalProbabilities-operators}
\title{\code{SurvivalProbabilities} operations}
\description{
A number of different operations can be performed on or with
\code{\link{SurvivalProbabilities}} objects. Methods have been defined for base
package generic operations including arithmetic and comparison.
}
\details{
The operations are:
\itemize{
\item \code{c}: concatenates a vector of \code{SurvivalProbabilities} objects
\item \code{[}: extract parts of a \code{SurvivalProbabilities} vector
\item \code{[<-}: replace parts of a \code{SurvivalProbabilities} vector
\item \code{rep}: repeat a \code{SurvivalProbabilities} object
\item \code{length}: determines the length of a \code{SurvivalProbabilities} vector
\item \code{*}: multiplication of \code{SurvivalProbabilities} objects. The end
date of the first SurvivalProbabilities object must be equivalent to the start
date of the second (or vice versa). Arguments are recycled as necessary.
\item \code{/}: division of \code{SurvivalProbabilities} objects. The start date
date of both arguments must be the same. Arguments are recycled as necessary.
\item \code{<, >, <=, >=, ==, !=}: these operate in the standard way on the
\code{discount_factor} field.
}
}
