% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{SurvivalProbabilities}
\alias{SurvivalProbabilities}
\title{Builds a \code{SurvivalProbabilitiesCurve}}
\usage{
SurvivalProbabilities(values, d1, d2, specs)
}
\arguments{
\item{specs}{CDS curve specifications that inherits from \code{\link[=CDSSpec]{CDSSpec()}}}

\item{survival_probabilities}{a vector of survival probabilities corresponding to each
time step in \code{tenors}.}
}
\value{
returns an object of type \code{SurvivalProbabilitiesCurve}
}
\description{
This will allow you to create a survival probability curve. This will
typically be bootstrapped from a \code{\link[=CDSCurve]{CDSCurve()}}.
}
\examples{











}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSMarkitSpec}},
  \code{\link{CDSSingleNameSpec}}, \code{\link{CDSSpec}},
  \code{\link{ZeroHazardRate}}, \code{\link{is.CDSCurve}},
  \code{\link{is.CDSSpec}}
}
\concept{CDS curve helpers}
