% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{VolSurface}
\alias{VolSurface}
\title{VolSurface class}
\usage{
VolSurface(reference_date, vol_quotes, ticker, surface_type, interpolation)
}
\arguments{
\item{reference_date}{a \code{Date} value that represents the as of date}

\item{vol_quotes}{a numeric matrix holding the vol values}

\item{ticker}{a label for the vol object, this should be the ticker of the underlying index}

\item{surface_type}{the type of the vol dimentions: "strike/term", "delta/term" or "moneyness/term"}

\item{interpolation}{a object of type interpolation \code{\link[fmbasics:Interpolation]{fmbasics::Interpolation()}}}
}
\description{
The `VolSurface`` class is designed to represent volatility surfaces.
Checks whether the dimentions of the surface correspond to the right type of axis: moneyness, delta or strike
}
