% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-env-class.R
\name{ZeroCurves}
\alias{ZeroCurves}
\title{Create a \code{ZeroCurves} pricing object}
\usage{
ZeroCurves(names, curves)
}
\arguments{
\item{names}{a character vector of curve labels with no duplicate values}

\item{curves}{a list of \link{ZeroCurve} each of which has the same
\code{reference_date}}
}
\value{
a \code{ZeroCurves} object that extends a tibble
}
\description{
A \code{ZeroCurves} object is effectively a \code{\link[tibble:tibble]{tibble::tibble()}} containing zero
curves.
}
\examples{
build_zero_curves()
}
\seealso{
Other pricing environments: \code{\link{FXRates}},
  \code{\link{PricingEnv}}, \code{\link{is.FXRates}},
  \code{\link{is.PricingEnv}}, \code{\link{is.ZeroCurves}},
  \code{\link{pick}}
}
