% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-class.R
\name{ZeroHazardRate}
\alias{ZeroHazardRate}
\title{Builds a \code{ZeroHazardRate}}
\usage{
ZeroHazardRate(value, compounding, day_basis, specs)
}
\arguments{
\item{specs}{CDS curve specifications that inherits from \code{\link[=CDSSpec]{CDSSpec()}}}

\item{hazard_rates}{a vector of hazard rates corresponding to each time step
in \code{tenors}}
}
\value{
returns an object of type \code{hazard_rates}
}
\description{
This will allow you to create a harzard rate curve. This will typically be
bootstrapped or implied from a \code{\link[=CDSCurve]{CDSCurve()}} or \code{\link[=SurvivalProbabilitiesCurve]{SurvivalProbabilitiesCurve()}}.
}
\examples{
c





}
\seealso{
Other CDS curve helpers: \code{\link{CDSCurve}},
  \code{\link{CDSMarkitSpec}},
  \code{\link{CDSSingleNameSpec}}, \code{\link{CDSSpec}},
  \code{\link{SurvivalProbabilities}},
  \code{\link{is.CDSCurve}}, \code{\link{is.CDSSpec}}
}
\concept{CDS curve helpers}
