% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmdata_example}
\alias{fmdata_example}
\title{Source supplied financial market data}
\usage{
fmdata_example(file)
}
\arguments{
\item{file}{the name of the file containing the data set.}
}
\value{
a \code{tibble}
}
\description{
The following files are supplied by the package:
\itemize{
\item \code{zerocurve.csv} has four fields for one curve: \code{start}, \code{end}, \code{zeros} and
\code{dfs} representing the start and end dates of the pillar instruments and
semi-annually compounded zero coupon rates and discount factors
}
}
\examples{
fmdata_example("zerocurve.csv")
}
\seealso{
\code{\link[=build_zero_curve]{build_zero_curve()}}
}
