% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate values from an object}
\usage{
interpolate(x, ...)
}
\arguments{
\item{x}{the object to interpolate.}

\item{...}{other parameters that defines how to interpolate the object}
}
\value{
an interpolated value or set of values
}
\description{
Interpolate values from an object
}
\seealso{
Other interpolate functions: \code{\link{interpolate.ZeroCurve}},
  \code{\link{interpolate_dfs}},
  \code{\link{interpolate_zeros}}
}
