% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-class.R
\name{interpolate.VolSurface}
\alias{interpolate.VolSurface}
\title{Interpolate a \code{VolSurface} object.}
\usage{
\method{interpolate}{VolSurface}(x, at, ...)
}
\arguments{
\item{x}{object of class \code{VolSurface} to be interpolated.}

\item{at}{indicates the coordinates at which the interpolation is performed.
\code{at} should be given as a \code{\link[tibble:tibble]{tibble::tibble()}} with two column names named
\code{maturity} and \code{smile}. e.g. list(maturity = c(1, 2), smile = c(72, 92)).}

\item{...}{unused in this model.}
}
\value{
\code{numeric} vector with length equal to the number of rows of \code{at}.
}
\description{
This method is used to interpolate a \code{VolSurface} object at multiple points of
the plane. The interpolation depends on the type of the surface, if the vols are
given by strikes, delta, moneyness.
}
\examples{
x <- build_vol_surface()
at <- tibble::tibble(
  maturity = c(as.Date("2020-03-31"), as.Date("2021-03-31")),
  smile = c(40, 80)
)
interpolate(x, at)
}
\seealso{
Other interpolate functions: \code{\link{interpolate.ZeroCurve}},
  \code{\link{interpolate_dfs}},
  \code{\link{interpolate_zeros}},
  \code{\link{interpolate}}
}
\concept{interpolate functions}
