% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R, R/vol-surface-class.R
\name{interpolate_vol}
\alias{interpolate_vol}
\alias{interpolate_vol.VolSurface}
\title{Interpolate volatility surface}
\usage{
interpolate_vol(x, maturity, strike, ...)

\method{interpolate_vol}{VolSurface}(x, maturity, strike, ...)
}
\arguments{
\item{x}{the object to interpolate}

\item{strike}{a vector of stikes to interpolate at with the same length as \code{date} vector}

\item{date}{a vetor of type date with the dates to interpolate at}
}
\value{
Returns a numeric vector of the same length as \code{date} parameter containing the interpolated
implied volatilities at each couple of \code{(date, strike)}.
}
\description{
This interpolates a vol surface from the class \link{VolsSurface} at a date and a strike
}
