% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-env-class.R
\name{is.FXRates}
\alias{is.FXRates}
\title{Inherits from FXRates}
\usage{
is.FXRates(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{FXRates} class; otherwise \code{FALSE}
}
\description{
Checks whether object inherits from \code{FXRates} class
}
\examples{
is.FXRates(build_fx_rates())
}
