% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/money-class.R
\name{MultiCurrencyMoney}
\alias{MultiCurrencyMoney}
\title{MultiCurrencyMoney}
\usage{
MultiCurrencyMoney(monies)
}
\arguments{
\item{monies}{a list of \link[=SingleCurrencyMoney]{SingleCurrencyMoney}}
}
\value{
a \code{MultiCurrencyMoney} object that extends \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This class associated a vector of numeric values with a list of currencies.
This can be useful for example to store value of cash flows.
}
\examples{
MultiCurrencyMoney(list(
  SingleCurrencyMoney(1, AUD()),
  SingleCurrencyMoney(2, USD())
))
}
\seealso{
Other money functions: \code{\link{CashFlow}},
  \code{\link{SingleCurrencyMoney}},
  \code{\link{is.CashFlow}},
  \code{\link{is.MultiCurrencyMoney}},
  \code{\link{is.SingleCurrencyMoney}}
}
