% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{VolSurface}
\alias{VolSurface}
\title{VolSurface class}
\usage{
VolSurface(vol_quotes, interpolation)
}
\arguments{
\item{vol_quotes}{object of class \code{\link[=VolQuotes]{VolQuotes()}} containing the volatility
data.}

\item{interpolation}{Interplation method, given as an object of class
interpolation \code{\link[=Interpolation]{Interpolation()}}. At this time only the
\code{\link[=LinearCubicTimeVarInterpolation]{LinearCubicTimeVarInterpolation()}} is supported.}
}
\value{
a \code{VolSurface} object
}
\description{
The \code{VolSurface} class is designed to capture the information about volatilities.
Checks whether the data members are of the correct type.
}
\examples{
build_vol_surface()
}
\seealso{
\link{interpolate.VolSurface}, \code{\link[=build_vol_surface]{build_vol_surface()}}
}
