% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{VolSurface}
\alias{VolSurface}
\title{VolSurface class}
\usage{
VolSurface(reference_date, surface_type, vol_quotes, dim1, dim2, ticker,
  interpolation)
}
\arguments{
\item{reference_date}{\code{Date} value that represents the as of date}

\item{surface_type}{type of the vol dimentions. Available types: "strike/term"}

\item{vol_quotes}{\code{matrix} of class \code{numeric} holding the vol values.}

\item{dim1}{\code{numeric} vector containing the surface strikes, deltas or
moneyness depending on the surface type.}

\item{dim2}{\code{Date} vector containing the tenors of the surface.}

\item{ticker}{label for the vol object, this should be the ticker of the underlying index}

\item{interpolation}{object of type interpolation \code{\link[fmbasics:Interpolation]{fmbasics::Interpolation()}}}
}
\description{
The `VolSurface`` class is designed to represent volatility surfaces.
Checks whether the dimentions of the surface correspond to the right type of axis: moneyness, delta or strike
}
