% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{VolSurface}
\alias{VolSurface}
\title{VolSurface class}
\usage{
VolSurface(reference_date, vol_quotes, ticker, interpolation)
}
\arguments{
\item{reference_date}{\code{Date} value that represents the as of date.}

\item{vol_quotes}{object of class \code{\link[=VolQuotes]{VolQuotes()}} containing the volatility data.}

\item{ticker}{a \code{string} label for the vol object.}

\item{interpolation}{Interplation method, given as an object of class
interpolation \code{\link[=Interpolation]{Interpolation()}}.}
}
\description{
The \code{VolSurface} class is designed to capture the information about volatilities.
Checks whether the data members are of the correct type.
}
\examples{
vol_surface <- build_vol_surface()
}
\seealso{
\link{interpolate.VolSurface}, \code{\link[=build_vol_surface]{build_vol_surface()}}
}
