% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-surface-class.R
\name{VolSurface}
\alias{VolSurface}
\title{VolSurface class}
\usage{
VolSurface(vol_quotes, interpolation)
}
\arguments{
\item{vol_quotes}{object of class \code{\link[=VolQuotes]{VolQuotes()}} containing the volatility
data.}

\item{interpolation}{Interplation method, given as an object of class
interpolation \code{\link[=Interpolation]{Interpolation()}}. At this time only
\code{\link[=LinearCubicTimeVarInterpolation]{LinearCubicTimeVarInterpolation()}} is supported. This is a two-dimensional
interpolator that uses linear interpolation in the time dimension and cubic
splines in the smile dimension with the values interpolated being the
square of the implied volatilities. Return values are implied volatilies}
}
\value{
a \code{VolSurface} object
}
\description{
The \code{VolSurface} class is designed to capture implied volatility information
along with information about how to interpolate an implied volatility between nodes.
}
\examples{
VolSurface(build_vol_quotes(), LinearCubicTimeVarInterpolation())
}
\seealso{
\link{interpolate.VolSurface}, \code{\link[=build_vol_surface]{build_vol_surface()}}
}
